Crate shell2batch

shell2batch

Coverts simple basic shell scripts to windows batch scripts.

While it is not really possible to take every shell script and automatically convert it to a windows batch file, this library provides a way to convert simple basic shell commands to windows batch commands.
The original goal of this library is to provide users of cargo-make a way to write simple tasks with shell scripts without duplicating their code for each platform.

Examples

extern crate shell2batch;

fn main() {
    let script = shell2batch::convert("rm ./myfile.txt");

    assert_eq!(script, "del ./myfile.txt");
    println!("Script: {}", script);
}

Installation

In order to use this library, just add it as a dependency:

[dependencies]
shell2batch = "*"

Contributing

See contributing guide

License

Developed by Sagie Gur-Ari and licensed under the Apache 2 open source license.

Functions

convert
Converts the provided shell script and returns the windows batch script text.